# reports/backend_implementation/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PDF Converter API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Document",
        "operationId": "document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/content": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Document Content",
        "operationId": "document_content_api_v1_documents__document_id__content_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job",
        "operationId": "job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Result",
        "operationId": "result_api_v1_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/downloads/text": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Download Text",
        "operationId": "download_text_api_v1_jobs__job_id__downloads_text_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/downloads/json": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Download Json",
        "operationId": "download_json_api_v1_jobs__job_id__downloads_json_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/assets/{asset_path}": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Asset",
        "operationId": "asset_api_v1_jobs__job_id__assets__asset_path__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "asset_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_post"
      },
...
      "ConversionCreate": {
        "properties": {
          "document_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Document Id"
          },
          "converter_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Converter Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "converter_id"
        ],
        "title": "ConversionCreate"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "content_url": {
            "type": "string",
            "title": "Content Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "size_bytes",
          "sha256",
          "page_count",
          "created_at",
          "expires_at",
          "content_url"
        ],
        "title": "DocumentResponse"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "converter_id": {
            "type": "string",
            "title": "Converter Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queued At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "processing_time_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Seconds"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "text_download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Download Url"
          },
          "json_download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Json Download Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "converter_id",
          "status",
          "stage",
          "created_at",
          "queued_at",
          "started_at",
          "finished_at",
          "processing_time_seconds",
          "warnings"
        ],
        "title": "JobResponse"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "validating",
          "processing",
          "standardizing",
          "exporting",
          "succeeded",
          "failed",
          "cancel_requested",
          "cancelled",
          "expired"
        ],
        "title": "JobStatus"
      },
      "Page": {
        "properties": {
          "page_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Number"
          },
          "width": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          },
          "text_blocks": {
            "items": {
              "$ref": "#/components/schemas/TextBlock"
            },
            "type": "array",
            "title": "Text Blocks"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/Table"
            },
            "type": "array",
            "title": "Tables"
          },
          "formulas": {
            "items": {
              "$ref": "#/components/schemas/Formula"
            },
            "type": "array",
            "title": "Formulas"
          },
          "chemical_objects": {
            "items": {
              "$ref": "#/components/schemas/ChemicalObject"
            },
            "type": "array",
            "title": "Chemical Objects"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageObject"
            },
            "type": "array",
            "title": "Images"
          },
          "diagrams": {
            "items": {
              "$ref": "#/components/schemas/DiagramObject"
            },
            "type": "array",
            "title": "Diagrams"
          },
          "reading_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reading Order"
          }
        },
        "type": "object",
        "required": [
          "page_number",
          "width",
          "height"
        ],
        "title": "Page"
      },
      "ResultResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "converter_id": {
            "type": "string",
            "title": "Converter Id"
          },
          "processing_time_seconds": {
            "type": "number",
            "title": "Processing Time Seconds"
          },
          "document": {
            "$ref": "#/components/schemas/StandardizedDocument"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "converter_id",
          "processing_time_seconds",
          "document"
        ],
        "title": "ResultResponse"
      },
      "StandardizedDocument": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "source_pdf": {
            "type": "string",
            "title": "Source Pdf"
          },
          "tool": {
            "$ref": "#/components/schemas/ToolMetadata"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/Page"
            },
            "type": "array",
            "title": "Pages"
          },
          "raw_artifacts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Raw Artifacts"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source_pdf",
          "tool"
        ],
        "title": "StandardizedDocument"
      },
...
      "TextBlock": {
        "properties": {
          "element_id": {
            "type": "string",
            "title": "Element Id"
          },
          "page_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Number"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text",
            "default": ""
          },
          "normalized_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Text"
          },
          "block_type": {
            "type": "string",
            "enum": [
              "paragraph",
              "heading",
              "list_item",
              "caption",
              "header",
              "footer",
              "unknown"
            ],
            "title": "Block Type",
            "default": "unknown"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          }
        },
        "type": "object",
        "required": [
          "element_id",
          "page_number"
        ],
        "title": "TextBlock"
      },
      "ToolMetadata": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "distribution_name": {
            "type": "string",
            "title": "Distribution Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "configuration": {
            "additionalProperties": true,
            "type": "object",
            "title": "Configuration"
          },
          "model_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Model Versions"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "distribution_name"
        ],
        "title": "ToolMetadata"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}